Move getStoryMarkup to output folder.
